If quantity is smaller than what is asked, disable allow

+++ src/Pages/Admin/Components/ItemRequest/ItemRequest.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState } from "react";
 import { Box, Button } from "@material-ui/core/";
 import { db } from "../../../../db";
 import "firebase/auth";
@@ -11,40 +11,54 @@ export default function ItemRequest({
   onRequestItem,
   onChangeOriginalItem,
 }) {
-  const handleAllow = () => {
-    db.collection("users")
-      .doc(user.id)
-      .collection("requesting")
-      .doc(item.id)
-      .delete()
-      .then(() => {
-        alert("Document successfully deleted!");
-        onRequestItem(itemRequests.filter((request) => request.id != item.id));
-      })
-      .catch((error) => {
-        alert("Error removing document: ", error);
-      });
+  const [canAllow, setCanAllow] = useState(true);
 
-    try {
-      db.collection("users").doc(user.id).collection("owned").doc(item.id).set({
-        name: item.name,
-        quantity: item.quantity,
-        id: item.id,
-      });
-      const originalItem = items.filter((list) => list.id == item.id)[0];
-      const editedQty = originalItem.quantity - item.quantity;
-      db.collection("items").doc(item.id).update({
-        quantity: editedQty,
-      });
-      const itemList = items.filter((original) => original.id != item.id);
-      const editedItem = {
-        id: item.id,
-        quantity: editedQty,
-        name: item.name,
-      };
-      onChangeOriginalItem([...itemList, editedItem]);
-    } catch (error) {
-      console.error(error);
+  const handleAllow = () => {
+    const originalItem = items.filter((list) => list.id == item.id)[0];
+    if (item.quantity < originalItem.quantity) {
+      db.collection("users")
+        .doc(user.id)
+        .collection("requesting")
+        .doc(item.id)
+        .delete()
+        .then(() => {
+          alert("Document successfully deleted!");
+          onRequestItem(
+            itemRequests.filter((request) => request.id != item.id)
+          );
+        })
+        .catch((error) => {
+          alert("Error removing document: ", error);
+        });
+
+      try {
+        db.collection("users")
+          .doc(user.id)
+          .collection("owned")
+          .doc(item.id)
+          .set({
+            name: item.name,
+            quantity: item.quantity,
+            id: item.id,
+          });
+        const editedQty = originalItem.quantity - item.quantity;
+        db.collection("items").doc(item.id).update({
+          quantity: editedQty,
+        });
+        const itemList = items.filter((original) => original.id != item.id);
+        const editedItem = {
+          id: item.id,
+          quantity: editedQty,
+          name: item.name,
+        };
+        onChangeOriginalItem([...itemList, editedItem]);
+      } catch (error) {
+        console.error(error);
+      }
+    } else {
+      console.log(`${originalItem.quantity} is less than ${item.quantity}`);
+      setCanAllow(false);
+      // Snackbar then immediate decline
     }
   };
   // TODO: modify quantity of item when allow
@@ -73,9 +87,11 @@ export default function ItemRequest({
         <div> {item.status} </div>
       ) : (
         <>
-          <Button variant="contained" color="secondary" onClick={handleAllow}>
-            allow
-          </Button>
+          {canAllow && (
+            <Button variant="contained" color="secondary" onClick={handleAllow}>
+              allow
+            </Button>
+          )}
 
           <Button variant="contained" color="secondary" onClick={handleDecline}>
             decline
